requirements.txt
streamlit==1.31.1
streamlit-option-menu==0.3.6
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
python-docx==0.8.11
PyPDF2==3.0.1
plotly==5.18.0
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0
requests==2.31.0